Initial Public ONI mod … POC, I guess? Metalama w/out Runtime
This is a snapshot of porting some of my personal mods to a public repo,
after I got metalama correctly doing code generation *without* a runtime
dependency from the mod to `Metalama.Framework`.

I was going to port a mod in too, but it is late, and that'll wait for
tomorrow, while the proof of concept can go out now. :)