Build: fix Publicise timestamp weirdness found in testing
...and probably only exists in testing, since it requires that the
timestamp of an input change without the content changing:

Publicize uses the SHA hash of the input file to determine if it should
do anything.  MSBuild uses the timestamp.

So, I could force the target to run incrementally by touch(1)-ing an
input file, but then Publicize did nothing, and the output file
timestamp never changed.

Which didn't break anything, but did cause Publicise to re-run every
time until the next patch changed the content of the non-publicised
assembly and all.

This fixes it by touch(1)-ing the modified files after running
Publicise, unconditionally, ensuring they are seens as up-to-date in
future build passes.