ModMain: remove forgotten `abstract` annotation...
Turns out I actually forgot to mark this as a concrete, not abstract,
class during past changes, and didn't actually test it until now.

Shockingly, everything worked once I removed that and recompiled, so
apparently that was sufficient to get everything working despite the
lack of real world testing to now!