MoreFoodCooksWhenHeated: my high temperature cooking mod
This is a "high temperature cooks food" mod, applying the Klei logic
that works on Raw Egg to Omelette to other things.  Being a "high
temperature cooking" mod there are many like it, but this one is mine.

It *does* bring one new feature to the table, which is
Codex integration: all the ingredients and food I modify have an
appropriate entry added to their conversions in the Codex showing that
they have the temperature transition.