MoreFoodCooksWhenHeated: fix unbalanced 1:1 conversions...
...or, "conversion", specifically, since Meat => CookedMeat was the only
thing on the list that had a mass ratio other than 1:1 for input
to output.

OTOH, a goal was that this should never be *better* than cooking things
manually, just avoiding using dupe time, so fixing it for that one
recipe is entirely appropriate as far as I'm concerned.

Also, updated the codex data to show the calorie and mass
transformation.  Not 100 percent sure it is the best possible
presentation of the data, but at least it is accurate and correct.