Build: disable VS2022 "Build Acceleration" since it breaks with ILRepack
Under some circumstances the Build Acceleration was determining that
the rebuilt SupportCode DLL could just be copied, and not triggering the
ILRepack step that it should.

So, disable it globally, and we can spend the ~ one second running
MSBuild instead of having a *second* build tool determine up-to-date
because our first one is a shitty horror that is too slow.