LogErrorNotifier: bring to public, use PLib Forwarding with it
I use a PLib PForwardedComponent to "elect" a winner among all the mods
running my LogErrorNotifier, and to route the logging hook through to
that single instance.

It also owns the LogErrorNotifier component, and attaching it to Game,
since the copies in each mod are technically different Type instances...

Anyway, it works, and because it simply passes a tiny interstitial
function to the other instances the relative cost should be low compared
to any other mechanism I could dream of inventing for sharing state.

Thanks, once again, PHan.  You still rock.