MoreFoodCooksWhenHeated: fix hook to run after recipes exist
I was hooking too early when I tried to process the recipes and get the
proportions: at that stage the multi-option recipes were registered, but
they had not yet generated the individual combinations.

So, almost all of 'em were missing.  Now, hooks at the right time, and
some bonus debugging for *why* things are not being added.